+++ packages/plugin-accounts-ui/src/components/account/detail/BasicInfo.tsx
@@ -1,10 +1,8 @@
-import Attachment from '@erxes/ui/src/components/Attachment';
 import Button from '@erxes/ui/src/components/Button';
 import DropdownToggle from '@erxes/ui/src/components/DropdownToggle';
 import Icon from '@erxes/ui/src/components/Icon';
 import ModalTrigger from '@erxes/ui/src/components/ModalTrigger';
 import { InfoWrapper } from '@erxes/ui/src/styles/main';
-import { IAttachment } from '@erxes/ui/src/types';
 import { __, Alert, confirm } from '@erxes/ui/src/utils';
 import { Action, Name } from '@erxes/ui-contacts/src/customers/styles';
 import Sidebar from '@erxes/ui/src/layout/components/Sidebar';
@@ -13,7 +11,7 @@ import {
   SidebarCounter,
   SidebarList
 } from '@erxes/ui/src/layout/styles';
-import AccountForm from '@erxes/ui-accounts/src/containers/AccountForm';
+import AccountForm from '../../../containers/ui_accounts/AccountForm';
 import { IAccount } from '../../../types';
 import React from 'react';
 import Dropdown from 'react-bootstrap/Dropdown';
