+++ packages/plugin-accounts-ui/src/components/account/detail/LeftSidebar.tsx
@@ -1,12 +1,10 @@
 import { gql } from '@apollo/client';
-import TaggerSection from '@erxes/ui-contacts/src/customers/components/common/TaggerSection';
 import Sidebar from '@erxes/ui/src/layout/components/Sidebar';
 import BasicInfo from '../../../containers/account/detail/BasicInfo';
 import CustomFieldsSection from '../../../containers/account/detail/CustomFieldsSection';
 import { IAccount } from '../../../types';
 import React from 'react';
 import { queries } from '../../../graphql';
-import { isEnabled } from '@erxes/ui/src/utils/core';
 
 type Props = {
   account: IAccount;
@@ -27,13 +25,6 @@ class LeftSidebar extends React.Component<Props> {
       <Sidebar wide={true}>
         <BasicInfo account={account} refetchQueries={refetchQueries} />
         <CustomFieldsSection account={account} />
-        {isEnabled('tags') && (
-          <TaggerSection
-            data={account}
-            type="accounts:account"
-            refetchQueries={refetchQueries}
-          />
-        )}
       </Sidebar>
     );
   }
