--- packages/ui-accounts/src/graphql/index.ts
@@ -1,4 +1,0 @@
-import mutations from './mutations';
-import queries from './queries';
-
-export { queries, mutations };
