--- packages/ui-accounts/src/index.ts
@@ -1,21 +1,0 @@
-import AccountChooser from './containers/AccountChooser';
-import AccountForm from './containers/AccountForm';
-import SelectAccounts from './containers/SelectAccounts';
-import CagetoryForm from './containers/CategoryForm';
-import {
-  queries as accountQueries,
-  mutations as accountMutations
-} from './graphql';
-import * as accountTypes from './types';
-import SelectAccountCategory from './containers/SelectAccountCategory';
-
-export {
-  CagetoryForm,
-  AccountChooser,
-  AccountForm,
-  SelectAccounts,
-  SelectAccountCategory,
-  accountQueries,
-  accountMutations,
-  accountTypes
-};
