--- packages/ui-accounts/src/styles.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const CategoryContainer = styled.div`
-  flex: 1;
-  flex-shrink: 0;
-`;
